ui/requirements.txt
werkzeug
#smbus
#sqlite3
python-memcached
flask
flask-sqlalchemy
flask-login
wtforms
pyserial
